+++ CompatibilityLib/base/src/main/java/com/elmakers/mine/bukkit/utility/platform/base/NBTUtilsBase.java
@@ -140,4 +140,9 @@ public abstract class NBTUtilsBase implements NBTUtils {
         }
         return result;
     }
+
+    @Override
+    public boolean setSpawnEggEntityData(ItemStack spawnEgg, Object entityData) {
+        return setTag(spawnEgg, "EntityTag", entityData);
+    }
 }