+++ Magic/src/main/java/com/elmakers/mine/bukkit/action/builtin/CaptureAction.java
@@ -84,7 +84,7 @@ public class CaptureAction extends BaseSpellAction {
         CompatibilityLib.getNBTUtils().removeMeta(savedEntity, "UUID");
         CompatibilityLib.getNBTUtils().removeMeta(savedEntity, "PortalCooldown");
 
-        if (!CompatibilityLib.getNBTUtils().setTag(spawnEgg, "EntityTag", savedEntity)) {
+        if (!CompatibilityLib.getNBTUtils().setSpawnEggEntityData(spawnEgg, savedEntity)) {
             return SpellResult.FAIL;
         }
         targetEntity.getWorld().dropItemNaturally(targetEntity.getLocation(), spawnEgg);